update tag type import
