+++ backend/core-api/src/modules/tags/@types/tag.ts
@@ -1,0 +1,9 @@
+import { IListParams } from 'core-api/@types';
+
+export interface ITagFilterQueryParams extends IListParams {
+  type: string;
+  tagIds?: string[];
+  parentId?: string;
+  ids: string[];
+  excludeIds: boolean;
+}
