+++ backend/libs/core-types/src/index.ts
@@ -1,8 +1,9 @@
 export * from './common';
-export * from './modules/contacts/customer';
+export * from './modules/app/app';
 export * from './modules/contacts/company';
 export * from './modules/contacts/contacts-common';
+export * from './modules/contacts/customer';
 export * from './modules/permissions/permission';
-export * from './modules/app/app';
-export * from './modules/team-member/user';
+export * from './modules/tags/tag';
 export * from './modules/team-member/structure';
+export * from './modules/team-member/user';
