+++ backend/libs/core-types/src/modules/tags/tag.ts
@@ -1,4 +1,3 @@
-import { IListParams } from 'core-api/@types';
 import { Document } from 'mongoose';
 
 export interface ITag {
@@ -14,12 +13,4 @@ export interface ITagDocument extends ITag, Document {
   createdAt: Date;
   order?: string;
   relatedIds?: string[];
-}
-
-export interface ITagFilterQueryParams extends IListParams {
-  type: string;
-  tagIds?: string[];
-  parentId?: string;
-  ids: string[];
-  excludeIds: boolean;
 }
